requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.3.0
